correct user deletion; renamed hello.py to app.py

+++ app.py
@@ -234,25 +234,30 @@ def get_current_date():
 
 # Delete Database Records
 @app.route("/delete/<int:id>")
+@login_required
 def delete(id):
-    user_to_delete = Users.query.get_or_404(id)
-    name = None
-    form = UserForm()
+    if id == current_user.id:
+        user_to_delete = Users.query.get_or_404(id)
+        name = None
+        form = UserForm()
 
-    try:
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        flash("User Deleted Successfully!")
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash("User Deleted Successfully!")
 
-        our_users = Users.query.order_by(Users.date_added)
-        return render_template(
-            "add_user.html", form=form, name=name, our_users=our_users
-        )
-    except:
-        flash("Whoops! There was an error deleting your user. Try again.")
-        return render_template(
-            "add_user.html", form=form, name=name, our_users=our_users
-        )
+            our_users = Users.query.order_by(Users.date_added)
+            return render_template(
+                "add_user.html", form=form, name=name, our_users=our_users
+            )
+        except:
+            flash("Whoops! There was an error deleting your user. Try again.")
+            return render_template(
+                "add_user.html", form=form, name=name, our_users=our_users
+            )
+    else:
+        flash("You cannot delete that user.")
+        return redirect(url_for("dashboard"))
 
 
 # Add User
